+++ Package.swift
@@ -54,7 +54,8 @@ let package = Package(
                                .product(name: "Atomics", package: "swift-atomics")]),
         .target(name: "ResourceExtension",
                 dependencies: ["OpenTelemetrySdk"],
-                path: "Sources/Instrumentation/sdk-resource-extension"),
+                path: "Sources/Instrumentation/SDKResourceExtension",
+                exclude: ["README.md"]),
         .target(name: "OpenTracingShim",
                 dependencies: ["OpenTelemetrySdk",
                                "Opentracing"]),
