+++ Sources/Instrumentation/SDKResourceExtension/ApplicationResourceProvider.swift
