+++ Sources/Instrumentation/SDKResourceExtension/DataSource/DeviceDataSource.swift
@@ -13,26 +13,25 @@
 // limitations under the License.
 //
 
-
-    import Foundation
+import Foundation
 #if os(watchOS)
     import WatchKit
 #elseif os(macOS)
-import AppKit
+    import AppKit
 #else
-import UIKit
+    import UIKit
 #endif
-    public class DeviceDataSource: IDeviceDataSource {
-        public var model: String? {
-            #if os(watchOS)
+public class DeviceDataSource: IDeviceDataSource {
+    public var model: String? {
+        #if os(watchOS)
             return WKInterfaceDevice.current().localizedModel
-            #else
+        #else
             let hwName = UnsafeMutablePointer<Int32>.allocate(capacity: 2)
             hwName[0] = CTL_HW
             #if os(macOS)
-            hwName[1] = HW_MODEL
+                hwName[1] = HW_MODEL
             #else
-            hwName[1] = HW_MACHINE
+                hwName[1] = HW_MACHINE
             #endif
             let machine = UnsafeMutablePointer<CChar>.allocate(capacity: 255)
             let len: UnsafeMutablePointer<Int>! = UnsafeMutablePointer<Int>.allocate(capacity: 1)
@@ -46,22 +45,21 @@ import UIKit
             }
             let machineName = String(cString: machine)
             return machineName
-            #endif
-        }
+        #endif
+    }
 
-        public var identifier: String? {
-            #if os(watchOS)
-            if #available (watchOS 6.3, *) {
+    public var identifier: String? {
+        #if os(watchOS)
+            if #available(watchOS 6.3, *) {
                 return WKInterfaceDevice.current().identifierForVendor?.uuidString
             } else {
                 return nil
             }
-            #elseif os(macOS)
+        #elseif os(macOS)
             return nil
-            #else
+        #else
             return UIDevice.current.identifierForVendor?.uuidString
 
-            #endif
-            
-        }
+        #endif
     }
+}
