+++ Sources/Instrumentation/SDKResourceExtension/DataSource/OperatingSystemDataSource.swift
@@ -1,0 +1,38 @@
+// Copyright 2021, OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+#if os(watchOS)
+    import WatchKit
+#elseif os(macOS)
+#else
+    import UIKit
+#endif
+
+import Foundation
+
+public class OperatingSystemDataSource: IOperatingSystemDataSource {
+    public var description: String {
+        ProcessInfo.processInfo.operatingSystemVersionString
+    }
+
+    public var type: String {
+        #if os(watchOS)
+            return "watchOS"
+        #elseif os(macOS)
+            return "macOS"
+        #else
+            return UIDevice.current.systemName
+        #endif
+    }
+}
