+++ Sources/Instrumentation/SDKResourceExtension/DefaultResources.swift
@@ -24,8 +24,8 @@ public class DefaultResources {
 
     let telemetry = TelemetryResourceProvider(source: TelemetryDataSource())
 
-    public init() {} 
-    
+    public init() {}
+
     public func get() -> Resource {
         var resource = Resource()
         let mirror = Mirror(reflecting: self)
